Rename seeding_data to more general sample_data.
