+++ databroker/examples/sample_data/synthetic_images.py
